Lighten RMCore class by moving recovery logic
